Analysis of leading domestic and foreign experience on higher education strategies in the context internationalization for sustainable social development: preprint (analytical materials) (Part I)
The presentation of the analytical materials outlines the theoretical foundations of higher education strategies in the context of internationalization for the sustainable development of society. The main attention is paid to the global institutional transformations, transformative strategies of higher education, the problem of quality, personally oriented strategies, multiculturalism, communication strategies of higher education.
The introduction of analytical materials will help to improve the level and quality of internationalization of higher education in Ukraine, to form an outlook for stimulating and ensuring sustainable development of society, enhancing the social responsibility of higher education and strengthening its function of serving the society.
For scholars in the field of philosophy and pedagogy, teachers, doctoral students, graduate students and students of higher education institutions.