Add files via upload
Internationalization is a global phenomenon, it is considered a priority for the development not only of the higher education system, but of society as a whole. That is why new models and approaches to the internationalization of higher education have emerged as a means of positioning the country beyond its borders.
The analytical materials provide an analysis of the strategies of internationalization of higher education in European countries: Italy, Germany, Poland, Finland, France and Russia.
Research on internationalization in these countries has illustrated differences in national and institutional strategies. However, common goals and objectives can be observed, such as reputation values (often symbolized by rating), visibility and competitiveness; competition for talented students and scholars; short-term and / or long-term economic benefits; demographic aspects; emphasis on international graduate employment and social activism.
An analysis of the internationalization of higher education in these countries has identified ten key trends: the increasing importance of internationalization at all levels; the impact of national internationalization strategies on institutional; insufficient funding for education; increasing the number of private institutions and the commercialization of educational services; the impact of globalization; growing competition; the growing role of regionalization; lack of data; new trends in internationalization (digital learning).
The results of the study can be used to develop a strategy for the internationalization of higher education in Ukraine.