Cultural and Humanitarian Strategies
In the monograph explores the internationalization of higher education as the main direction of its development in the context of dynamic social transformations caused by globalization and the rapid development of the latest information and computer technologies. The personal and institutional dimensions of the internationalization of higher education, its multicultural and transformative potential, in particular in the cosmopolitan perspective, the mechanisms of internationalization of the universities of the European Union and the communication processes in the internationalization of higher education are analyzed. It is substantiated that the internationalization of higher education should be considered not as a self-sufficient goal, but as a means of improving the quality of education and a means of preparing a young person for life in a globalized world in the conditions of dynamic civilizational transformations. For this purpose, the internationalization of higher education must be culturally-humanitarian and implemented using a personal approach. The monograph analyzes strategic documents of the global, European and national levels that determine the policy and perspectives of internationalization of higher education.