Internationalization of Higher Education
Methodological materials can be used by the Ministry of Education and Science of Ukraine, institutions of higher education of Ukraine, institutions of National Academy of Sciences of Ukraine, the councils of rectors of regional university centers of Ukraine and positively influence the process of reforming university education in accordance with modern world standards. The materials and results of the analytical review may also be of interest to researchers, teachers, young scientists, masters, students, as well as all those interested in issues of internationalization of higher education and its cultural and humanitarian strategies.